requirements.txt
requests
python-dotenv
fastapi
uvicorn
beautifulsoup4
readability-lxml
lxml